Update specs.py
change np.int32 to np.int64

+++ specs.py
@@ -45,7 +45,7 @@ def get_observation_action_spec_tabular(bounds):
 
     return get_observation_action_spec(
         obs_shape=(), act_shape=(),
-        obs_dtype=np.int32, act_dtype=np.int32,
+        obs_dtype=np.int64, act_dtype=np.int64,
         obs_min=obs_min, obs_max=obs_max, act_min=act_min, act_max=act_max,
     )
 
@@ -55,7 +55,7 @@ def get_observation_action_spec_continuous(bounds, shapes):
 
     return get_observation_action_spec(
         obs_shape=obs_shape, act_shape=act_shape,
-        obs_dtype=np.float32, act_dtype=np.int32,
+        obs_dtype=np.float32, act_dtype=np.int64,
         obs_min=obs_min, obs_max=obs_max, act_min=act_min, act_max=act_max,
     )
